+++ conversejs/loc.keys.js
@@ -14,6 +14,7 @@ const locKeys = [
   'login_remote_peertube_video_not_found',
   'login_remote_peertube_video_not_found_try_anyway',
   'login_remote_peertube_video_not_found_try_anyway_button',
+  'login_remote_peertube_video_open_failed',
   'login_external_auth_alert_message'
 ]